LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-async-respond-with.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancelled

Harness Error (FAIL), message = cancelled

PASS global setup 
PASS respondWith in a task throws InvalidStateError 
PASS respondWith in a microtask does not throw 
PASS global cleanup 
